Defensive programming added to stop unauthorized access to a recipe that is not owned by the creator

+++ app.py
@@ -239,6 +239,16 @@ def edit_recipe(recipe_id):
 # Route for delete recipe
 @app.route("/delete_recipe/<recipe_id>")
 def delete_recipe(recipe_id):
+    if "user" not in session:
+        flash("Please log in or register to see this page")
+        return redirect(url_for("login"))
+
+    recipe = mongo.db.recipes.find_one({"_id": ObjectId(recipe_id)})
+
+    if recipe is None or recipe.get('author_id') != session["user"]:
+        flash("You are not authorized to access this recipe")
+        return redirect(url_for("get_recipes"))
+        
     mongo.db.recipes.delete_one({"_id": ObjectId(recipe_id)})
     flash("Your recipe has been deleted")
     return redirect(url_for("get_recipes"))
